alx_travel_app/requirement.txt
Django>=4.2
djangorestframework
django-cors-headers
celery
drf-yasg
mysqlclient
django-environ

